# docs/openapi.yaml
openapi: 3.0.3
info:
  title: FinTrack API
  description: >
    Production REST API for FinTrack Personal Financial Management and Asset Tracking.
    Implements cash flow tracking, dynamic budgeting, income allocation, customizable
    financial cycles, receivables (loans), investments, and periodic reporting.
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Local Docker Compose Environment
paths:
...
  /cycle/current:
    get:
      summary: Get Current Cycle Window
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Active cycle boundaries
  /cycle/summary:
    get:
      summary: Get Cycle Cash Flow Summary
      security:
        - BearerAuth: []
      parameters:
        - name: date
          in: query
          schema: { type: string, format: date }
      responses:
        '200':
          description: Cycle operational metrics
  /budgets:
    get:
      summary: List Budgets
      security:
        - BearerAuth: []
      parameters:
        - name: cycle_start
          in: query
          schema: { type: string, format: date }
        - name: cycle_end
          in: query
          schema: { type: string, format: date }
      responses:
        '200':
          description: Budget limits with actual spend and variance
    post:
      summary: Create Category Budget
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [category_id, cycle_start, cycle_end, planned_amount]
              properties:
                category_id: { type: string, format: uuid }
                cycle_start: { type: string, format: date }
                cycle_end: { type: string, format: date }
                planned_amount: { type: integer, minimum: 0 }
      responses:
        '201':
          description: Budget limit created
...
  /investments:
    get:
      summary: List Investments
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Portfolio holdings with capital and valuation
    post:
      summary: Add Investment Holding
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, type, capital]
              properties:
                name: { type: string }
                type: { type: string, enum: [STOCK, MUTUAL_FUND, GOLD, CRYPTO, OTHER] }
                capital: { type: integer, minimum: 0 }
                current_value: { type: integer, minimum: 0 }
                source_account_id: { type: string, format: uuid }
                notes: { type: string }
      responses:
        '201':
          description: Holding created
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT